@@ -1,7 +1,13 @@
 {
   "name": "vxweb",
+  "scripts": {
+    "vue-backend-build": "vue-cli-service build --no-clean --target lib --dest 'dist/vue' --name vxweb vue/build/vxweb.js",
+    "vue-sample-build": "vue-cli-service build --no-clean --target lib --dest 'dist/vue' --name sample vue/build/sample.js"
+  },
   "devDependencies": {
-    "@vue/cli-service": "^4.5.9",
+    "@vue/cli-service": "^5.0.0-alpha.8",
+    "@vue/compiler-sfc": "^3.0.11",
+    "browserslist": "^4.16.3",
     "gulp": "latest",
     "gulp-autoprefixer": "latest",
     "gulp-clean-css": "^4.3.0",
@@ -9,23 +15,18 @@
     "gulp-rename": "^2.0.0",
     "gulp-sass": "^4.1.0",
     "node-sass-tilde-importer": "^1.0.2",
-    "sass-loader": "^10.1.0",
+    "sass": "^1.32.10",
+    "sass-loader": "^10.1.1",
     "spectre.css": "^0.5.9",
-    "vue": "^2.6.12",
-    "vue-slicksort": "^1.1.3",
-    "vue-template-compiler": "^2.6.12"
+    "style-resources-loader": "^1.4.1",
+    "vue": "^3.0.11",
+    "vue-cli-plugin-style-resources-loader": "~0.1.5",
+    "vue-slicksort": "^1.1.3"
   },
   "browserslist": [
     "last 3 chrome versions",
     "last 3 edge versions",
     "firefox esr",
     "last 3 safari versions"
-  ],
-  "scripts": {
-    "vue-backend-build": "vue-cli-service build --no-clean --target lib --dest 'dist/vue' --name vxweb vue/build/vxweb.js",
-    "vue-sample-build": "vue-cli-service build --no-clean --target lib --dest 'dist/vue' --name sample vue/build/sample.js"
-  },
-  "dependencies": {
-    "browserslist": "^4.16.3"
-  }
+  ]
 }
