@@ -2,9 +2,6 @@
     <input
         v-bind="$attrs"
         class="form-input"
-        v-on="{
-          ...$listeners,
-          input: event => $emit('input', event.target.value)
-        }"
+        @input="event => $emit('input', event.target.value)"
     >
 </template>
